+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 1.2.0
+
+- Fixed decorations not being applied correctly when the itemCount is 1
+- Added `createVertical` and `createHorizontal` helpers to:
+    - `LinearMarginDecoration`
+    - `LinearBoundsMarginDecoration`
+    - `GridBoundsMarginDecoration
+
 # 1.1.0
 
 - Added `GridSpanBoundsMarginDecoration` to apply margins to the bounds of RecyclerViews that use a GridLayoutManager with different span sizes